# swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Urbano 2
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /contactos:
    x-swagger-router-controller: contacto
    get:
      tags: 
        - Contactos
      operationId: listarContactos
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Contacto'
    post:
      tags: 
        - Contactos
      operationId: crearContacto
      parameters:
        - in: body
          name: body
          description: Datos del usuario a crear.
          schema:
            type: object
            required:
              - nombre
              - apellidos
            properties:
              nombre:
                type: string
              segundoNombre:
                type: string
              sexo:
                type: string
              enorganigrama:
                type: ['string','null']
              cargo:
                type: ['string','null']
              area:
                type: string
              apellidos:
                type: string
              servicios:
                type: array
                items:
                  type: object
                  properties:
                    nombre:
                      type: string
                    pin:
                      type: string
                    checked:
                      type: boolean
              telefono:
                type: array
                items:
                  type: object
                  properties:
                    nombre:
                      type: string
                    valor:
                      type: string
                    checked:
                      type: boolean
              correos:
                type: array
                items:
                  type: object
                  properties:
                    nombre:
                      type: string
                    valor: 
                      type: string
              direcciones:
                type: array
                items:
                  type: object
              roles:
                type: array
                items:
                  type: object
                  properties:
                    nombre:
                      type: string
                    valor: 
                      type: string
                    checkedAgenda:
                      type: boolean
      responses:
        200:
          description: OK
...
  /contactos/{id}:
    x-swagger-router-controller: contacto
    get:
      tags: 
        - Contactos
      operationId: listarContactoPorId
      parameters:
        - name: id
          in: path
          description: ID del usuario
          type: string
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Contacto'
    put:
      tags: 
        - Contactos
      operationId: editarContacto
      parameters:
        - name: id
          in: path
          description: ID del contacto
          type: string
          required: true
        - in: body
          name: body
          description: Actualizar contacto por ID
          schema:
            type: object
            required:
              - nombre
              - apellidos
            properties:
              nombre:
                type: string
              segundoNombre:
                type: string
              sexo:
                type: string
              enorganigrama:
                type: ['string','null']
              cargo:
                type: ['string','null']
              area:
                type: string
              apellidos:
                type: string
              servicios:
                type: array
                items:
                  type: object
                  properties:
                    nombre:
                      type: string
                    pin:
                      type: string
                    checked:
                      type: boolean
              telefono:
                type: array
                items:
                  type: object
                  properties:
                    nombre:
                      type: string
                    valor:
                      type: string
                    checked:
                      type: boolean
              correos:
                type: array
                items:
                  type: object
                  properties:
                    nombre:
                      type: string
                    valor: 
                      type: string
              direcciones:
                type: array
                items:
                  type: object
              roles:
                type: array
                items:
                  type: object
                  properties:
                    nombre:
                      type: string
                    valor: 
                      type: string
                    checkedAgenda:
                      type: boolean
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Contacto'
    delete:
      tags: 
        - Contactos
      operationId: eliminarContacto
      parameters:
        - name: id
          in: path
          description: ID del contacto
          type: string
          required: true
      responses:
        200:
          description: OK
  # //////////////////////////////////////////////////////////////////// Organigrama
...
definitions:
  Contacto:
    type: object
    properties:
      id:
        type: string
      nombre:
        type: string
      segundoNombre:
        type: string
      sexo:
        type: string
      enorganigrama:
        type: string
      cargo:
        type: string
      area:
        type: string
      apellidos:
        type: string
      servicios:
        type: array
        items:
          type: object
          properties:
            nombre:
              type: string
            pin:
              type: string
            checked:
              type: boolean
      telefono:
        type: array
        items:
          type: object
          properties:
            nombre:
              type: string
            valor:
              type: string
            checked:
              type: boolean
      correos:
        type: array
        items:
          type: object
          properties:
            nombre:
              type: string
            valor: 
              type: string
      direcciones:
        type: array
        items:
          type: object
      roles:
        type: array
        items:
          type: object
          properties:
            nombre:
              type: string
            valor: 
              type: string
            checkedAgenda:
              type: boolean
...